[TAN-2161] Preserve old behavior of raising error in cache_tenant if no current tenant
