+++ back/app/controllers/application_controller.rb
@@ -70,8 +70,8 @@ class ApplicationController < ActionController::API
   # Used by semantic logger to include in every log line
   def append_info_to_payload(payload)
     super
-    payload[:tenant_id] = Current.tenant&.id
-    payload[:tenant_host] = Current.tenant&.host
+    payload[:tenant_id] = Tenant.safe_current&.id
+    payload[:tenant_host] = Tenant.safe_current&.host
     payload[:user_id] = current_user&.id
     payload[:request_id] = request.request_id
     payload[:'X-Amzn-Trace-Id'] = request.headers['X-Amzn-Trace-Id']
