+++ back/engines/commercial/multi_tenancy/app/controllers/multi_tenancy/patches/application_controller.rb
@@ -11,7 +11,7 @@ module MultiTenancy
       end
 
       def set_sentry_context
-        Sentry.set_tags(tenant: Current.tenant&.host)
+        Sentry.set_tags(tenant: Tenant.safe_current&.host)
         Sentry.set_user(id: current_user.id) if current_user
       end
